Fixed IMU offset by adding accelerometer offset term

+++ src/ADXL345/ADXL345.cpp
@@ -507,7 +507,7 @@ AccelG ADXL345::getAccelG(void)
     } 
     else 
     {   
-        scale = (float) range / 1024;
+        scale = range / 1024.0;
     }
     
     AccelG res;
